+++ test/unit/test-serialize-js-utils.js
@@ -1,3 +1,0 @@
-/**
- * Created by jens.krefeldt on 24.08.17.
- */